couch-doc-diff: diff of a document on the couch (feature 3485)
Added the bash script couch-doc-diff.
It runs a diff between the content of the couch document specified as the first parameter and the second parameter.
Diff returns 0 if there is no difference. This way you can tell the data is already on the couch.
The diff will ignore changes to whitespaces as these are usually irrelevant for security json docs.